add navigation prototype
currently viewmodels have access to a class holding the backStack of
composables wrapped in Screen interface, this might require improvement
down the line so the business logic is less aware of how are things
displayed.

On the potentially plus side, all navigation decision logic takes place
in the viewmodel, even back press has to be go through the VM, same for
any other transition. Granted it depends on developer adherence to the
convention though.