separate buisness logic from view in navigation
Navigator no longer knows about Screens, instead handles
NavDestinations, one of which is Previous. Instead main function
interprets which destination leads to which screen and holds a backStack
of the visited screens.

This might still need some improvements, like a facade to handle
backStack for developer and only ask to interpret the destinations.