fix typing screen failing when database is empty
When database is empty, and user types something then deletes it, it was
taken as a correct answer since "" string was set when there was no
challenge word active.

Also fixed an exception when random() was called on an empty collection
and improved UI behavior during the test by ensuring that Text for
results always exists, rather than appearing only when score is ready
and pushing input field upwards, as they both are arranged in the middle
of the screen.